+++ tests/test_templates.py
@@ -39,6 +39,32 @@ def test_templates(tmpdir, test_client_factory, apostrophe):
     assert set(response.context.keys()) == {"request"}
 
 
+def test_async_templates(tmpdir, test_client_factory):
+    path = os.path.join(tmpdir, "index.html")
+    with open(path, "w") as file:
+        file.write("<html>Hello {{ async_fn() }}</html>")
+
+    async def async_fn():
+        return "world"
+
+    async def homepage(request):
+        return templates.get_template_response(
+            request, "index.html", context={"async_fn": async_fn}
+        )
+
+    app = Lilya(
+        debug=True,
+        routes=[Path("/", handler=homepage)],
+    )
+    templates: Jinja2Template = Jinja2Template(directory=str(tmpdir), enable_async=True)
+
+    client = test_client_factory(app)
+    response = client.get("/")
+    assert response.text == "<html>Hello world</html>"
+    assert response.template.name == "index.html"
+    assert set(response.context.keys()) == {"request", "async_fn"}
+
+
 def test_calls_context_processors(tmp_path, test_client_factory):
     path = tmp_path / "index.html"
     path.write_text("<html>Hello {{ username }}</html>")